withdrawals

+++ internal/models/session/session.go
@@ -9,5 +9,5 @@ import (
 type Session struct {
 	ID        uuid.UUID `json:"id"`
 	CreatedAt time.Time
-	UserID    int64
+	UserID    int
 }
