+++ internal/models/user/user.go
@@ -6,7 +6,7 @@ import (
 )
 
 type User struct {
-	ID             int64  `json:"id"`
+	ID             int    `json:"id"`
 	Login          string `json:"login"`
 	HashedPassword string `json:"-"`
 	Salt           string
